long tests 1 (math, eq,neq) passed, test 2 (cmp) failed

+++ include/vm/stacks.h
@@ -49,7 +49,7 @@ void pvm_istack_reserve( struct data_area_4_integer_stack* rootda, int n_slots )
 
 void 				pvm_lstack_push( struct data_area_4_integer_stack* rootda, int64_t o );
 int64_t				pvm_lstack_pop( struct data_area_4_integer_stack* rootda );
-int 				pvm_lstack_top( struct data_area_4_integer_stack* rootda );
+int64_t				pvm_lstack_top( struct data_area_4_integer_stack* rootda );
